+++ src/iana/requested/cose.ts
@@ -25,8 +25,16 @@ export enum draft_headers {
   receipts = 394,
   verifiable_data_structure = 395,
   verifiable_data_proofs = 395
+}
 
+export const VerifiableDataStructures = {
+  'RFC9162-Binary-Merkle-Tree': 1
+}
 
+// only for RFC9162-Binary-Merkle-Tree
+export const VerifiableDataProofTypes = {
+  'RFC9162-Inclusion-Proof': -1,
+  'RFC9162-Consistency-Proof': -2
 }
 
 export { algorithms_to_labels, labels_to_algorithms }
